refactor(avatar): share Composer mascot and restore energy effects
