+++ desktop/src/renderer/components/conversation/styles/composer-focus.css
@@ -29,75 +29,7 @@
   z-index: 2;
 }
 
-@keyframes composer-mascot-launch {
-  0% {
-    transform: translateY(0) scale(1, 1);
-  }
-  14% {
-    transform: translateY(3px) scale(1.14, 0.8);
-  }
-  38% {
-    transform: translateY(-20px) scale(0.9, 1.14);
-  }
-  58% {
-    transform: translateY(0) scale(1.1, 0.88);
-  }
-  74% {
-    transform: translateY(-6px) scale(0.98, 1.04);
-  }
-  88% {
-    transform: translateY(0) scale(1.02, 0.98);
-  }
-  100% {
-    transform: translateY(0) scale(1, 1);
-  }
-}
 
-.composer-mascot-launch {
-  animation: composer-mascot-launch 820ms cubic-bezier(0.3, 0.9, 0.4, 1);
-  transform-origin: bottom center;
-}
-
-@keyframes composer-mascot-jelly {
-  0% {
-    transform: scale(1, 1);
-  }
-  25% {
-    transform: scale(1.16, 0.84);
-  }
-  45% {
-    transform: scale(0.9, 1.1);
-  }
-  62% {
-    transform: scale(1.06, 0.94);
-  }
-  78% {
-    transform: scale(0.98, 1.02);
-  }
-  100% {
-    transform: scale(1, 1);
-  }
-}
-
-.composer-mascot-jelly {
-  transform-origin: bottom center;
-}
-
-.composer-mascot-jelly:hover {
-  animation: composer-mascot-jelly 600ms ease-out;
-}
-
-@keyframes composer-mascot-bubble-in {
-  0% {
-    opacity: 0;
-    transform: translateY(6px) scale(0.94);
-  }
-  100% {
-    opacity: 1;
-    transform: translateY(0) scale(1);
-  }
-}
-
-.composer-mascot-bubble {
-  animation: composer-mascot-bubble-in 260ms cubic-bezier(0.34, 1.56, 0.64, 1);
+.composer-mascot-character-stage > [data-dotcraft-plugin-surface='composer.mascot'] > * {
+  grid-area: 1 / 1;
 }
