+++ desktop/src/renderer/styles/index.css
@@ -5,14 +5,8 @@
 @import "./foundations/themes.css";
 @import "./foundations/base.css";
 @import "./foundations/code-tokens.css";
-@import "../components/conversation/styles/mascot/foundation.css";
-@import "../components/conversation/styles/mascot/overdrive.css";
-@import "../components/conversation/styles/mascot/idle.css";
 @import "../components/conversation/styles/composer-focus.css";
 @import "../components/conversation/styles/composer-submit.css";
-@import "../components/conversation/styles/mascot/expressions.css";
-@import "../components/conversation/styles/mascot/feedback.css";
-@import "../components/conversation/styles/mascot/props.css";
 @import "../components/conversation/styles/reference-chips.css";
 @import "../components/conversation/styles/message-stream.css";
 @import "../components/conversation/styles/tool-running.css";
