+++ desktop/src/renderer/tests/ApprovalDecisionComposer.test.tsx
@@ -90,7 +90,7 @@ describe('ApprovalDecisionComposer', () => {
     expect(mascot).toHaveAttribute('data-mascot-effort', 'high')
     expect(mascot).toHaveAttribute('data-mascot-speed', 'fast')
     expect(mascot).toHaveAttribute('data-mascot-context', 'max')
-    expect(mascot).toHaveClass('composer-mascot-hold-sign')
+    expect(mascot?.querySelector('.dca-robot')).toHaveAttribute('data-pose', 'waiting')
   })
 
   it('keeps subscription status independent when workspace status is replaced', () => {
