+++ desktop/src/renderer/tests/RequestUserInputComposer.test.tsx
@@ -140,7 +140,7 @@ describe('RequestUserInputComposer', () => {
     expect(mascot).toHaveAttribute('data-mascot-effort', 'high')
     expect(mascot).toHaveAttribute('data-mascot-speed', 'fast')
     expect(mascot).toHaveAttribute('data-mascot-context', 'max')
-    expect(mascot).toHaveClass('composer-mascot-hold-sign')
+    expect(mascot?.querySelector('.dca-robot')).toHaveAttribute('data-pose', 'waiting')
   })
 
   it('renders in ConversationPanel instead of the normal composer', async () => {
