+++ sdk/typescript/packages/avatar/src/composer/styles/overdrive.css
@@ -1,5 +1,3 @@
-/* MAX soft overdrive: the fixed values are the 200% outline setting from the design
-   preview. */
 [data-mascot-context='max'] .composer-mascot-motion {
   --mascot-max-edge: oklch(from var(--mascot-energy-accent)
     clamp(0.28, calc(l + 0.065), 0.96) clamp(0, calc(c * 0.96), 0.3) calc(h + 4));
@@ -107,22 +105,6 @@
   68% { opacity: 0.13; transform: translate(-48%, -54%) scale(1.04); }
 }
 
-@keyframes composer-mascot-max-arm-right {
-  0%, 50%, 94%, 100% { color: var(--mascot-body-light); }
-  72% {
-    color: oklch(from var(--mascot-body-light)
-      clamp(0.28, calc(l + 0.15), 0.96) clamp(0, calc(c * 0.58), 0.3) calc(h + 41));
-  }
-}
-
-@keyframes composer-mascot-max-arm-left {
-  0%, 34%, 100% { color: var(--mascot-body-dark); }
-  14% {
-    color: oklch(from var(--mascot-body-dark)
-      clamp(0.28, calc(l + 0.09), 0.96) clamp(0, calc(c * 0.9), 0.3) calc(h + 14));
-  }
-}
-
 [data-mascot-profile-transition='active'] .composer-mascot-character-stage {
   transform-origin: 50% 52%;
   animation: composer-mascot-profile-character 1.24s linear both;
@@ -200,12 +182,12 @@
     transform: translateY(0) rotate(1.2deg) scale(0.76, 0.8);
     filter: brightness(0.96) saturate(1.08);
   }
-  48% {
+  46% {
     opacity: 0.14;
     transform: translateY(0) rotate(-4deg) scale(0.13);
     filter: brightness(0.72) saturate(1.18) blur(0.7px);
   }
-  49.5%, 50.5% {
+  48%, 52% {
     opacity: 0;
     transform: translateY(0) rotate(-6deg) scale(0.02);
     filter: brightness(0.58) saturate(1.2) blur(1px);
