+++ sdk/typescript/packages/avatar/src/composer/styles/props.css
@@ -16,3 +16,6 @@
     transition: none !important;
   }
 }
+
+[data-composer-mascot-motion='off'],
+[data-composer-mascot-motion='off'] * { animation: none !important; transition: none !important; }
