+++ sdk/typescript/packages/avatar/src/react.ts
@@ -1,2 +1,6 @@
 export { Avatar, AppearanceAvatar, type AvatarProps, type AppearanceAvatarProps } from './Avatar.js'
 export { DecorationSwatch, PrimaryDecoration, SecondaryDecoration } from './Decorations.js'
+
+export { ComposerMascot } from './composer/ComposerMascot.js'
+export type { ComposerMascotProps, ComposerMascotContext, MascotExpression, MascotLight } from './composer/types.js'
+export { ComposerMascotShadow } from './composer/ComposerMascotShadow.js'
