+++ sdk/typescript/packages/avatar/src/styles.css
@@ -1,3 +1,13 @@
 @import './rig.css';
 @import './avatar.css';
 @import './accessories.css';
+
+@import "./composer/styles/foundation.css";
+@import "./composer/styles/overdrive.css";
+@import "./composer/styles/idle.css";
+@import "./composer/styles/focus.css";
+@import "./composer/styles/expressions.css";
+@import "./composer/styles/feedback.css";
+@import "./composer/styles/props.css";
+
+@import "./composer/styles/energy.css";
